+++ src/renderer/components/SongList/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import List from 'material-ui/List';
 
-import Song, { SongData } from '../../components/Song';
+import Song, { SongData } from 'components/Song';
 
 interface IProps extends React.Props<SongList> {
   songs: SongData[];
