+++ src/renderer/containers/App/index.tsx
@@ -2,11 +2,11 @@ import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import { MuiThemeProvider } from 'material-ui/styles';
 
-import * as colors from '../../colors';
-import NavBar from '../../components/NavBar';
-import Player from '../../components/Player';
-import HomePage from '../HomePage';
-import Settings from '../Settings';
+import * as colors from 'utils/colors';
+import NavBar from 'components/NavBar';
+import Player from 'components/Player';
+import HomePage from 'containers/HomePage';
+import Settings from 'containers/Settings';
 import { PageContainer } from './styles';
 
 export default class App extends React.Component<{}, {}> {
